# desktop/openapi.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "Quater Water Quality Management API",
    "description": "REST API for managing water quality testing data, compliance calculations, and laboratory operations",
    "contact": {
      "name": "Quater Development Team",
      "email": "support@quater.app"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/Auth/token": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "OAuth2 token endpoint - handles authorization_code and refresh_token grant types.\n\nFor authorization code exchange:\n  POST /api/auth/token\n  Content-Type: application/x-www-form-urlencoded\n  \n  grant_type=authorization_code&code=AUTH_CODE&code_verifier=PKCE_VERIFIER&redirect_uri=quater://oauth/callback&client_id=quater-mobile\n\nFor refresh:\n  POST /api/auth/token\n  Content-Type: application/x-www-form-urlencoded\n  \n  grant_type=refresh_token&refresh_token=YOUR_REFRESH_TOKEN",
        "parameters": [
          {
            "name": "api-version",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Auth/logout": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "Logout and revoke all tokens for the user",
        "description": "This endpoint revokes ALL tokens for the user, effectively logging them out from all devices.\nFor per-device logout, use the /revoke endpoint with the specific refresh token.",
        "parameters": [
          {
            "name": "api-version",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/Auth/userinfo": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "Get current user information",
        "parameters": [
          {
            "name": "api-version",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
...
      "UserDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userName": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "labs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UserLabDto"
            },
            "nullable": true
          },
          "lastLogin": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "isActive": {
            "type": "boolean"
          }
        },
        "additionalProperties": false
      },
...
      "UserLabDto": {
        "required": [
          "labName",
          "role"
        ],
        "type": "object",
        "properties": {
          "labId": {
            "type": "string",
            "format": "uuid"
          },
          "labName": {
            "maxLength": 200,
            "minLength": 1,
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/UserRole"
          },
          "assignedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "UserRole": {
        "enum": [
          1,
          2,
          3
        ],
        "type": "integer",
        "format": "int32"
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT Authorization header using the Bearer scheme. Enter 'Bearer' [space] and then your token in the text input below.",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    { }
  ],
...
}